+++ android/src/main/kotlin/dev/msfjarvis/claw/android/paging/SearchPagingSource.kt
@@ -15,16 +15,13 @@ import dagger.assisted.AssistedInject
 import dev.msfjarvis.claw.android.paging.LobstersPagingSource.Companion.PAGE_SIZE
 import dev.msfjarvis.claw.android.paging.LobstersPagingSource.Companion.STARTING_PAGE_INDEX
 import dev.msfjarvis.claw.android.ui.toError
-import dev.msfjarvis.claw.android.viewmodel.ReadPostsRepository
-import dev.msfjarvis.claw.android.viewmodel.SavedPostsRepository
 import dev.msfjarvis.claw.api.LobstersSearchApi
 import dev.msfjarvis.claw.core.injection.IODispatcher
-import dev.msfjarvis.claw.database.local.SavedPost
+import dev.msfjarvis.claw.model.LobstersPost
 import dev.msfjarvis.claw.model.UIPost
 import dev.msfjarvis.claw.model.toUIPost
 import java.io.IOException
 import kotlinx.coroutines.CoroutineDispatcher
-import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.withContext
 
 /**
@@ -39,19 +36,8 @@ constructor(
   private val searchApi: LobstersSearchApi,
   @Assisted private val queryProvider: () -> String,
   @IODispatcher private val ioDispatcher: CoroutineDispatcher,
-  private val savedPostsRepository: SavedPostsRepository,
-  private val readPostsRepository: ReadPostsRepository,
 ) : PagingSource<Int, UIPost>() {
-  private lateinit var savedPosts: List<String>
-  private lateinit var readPosts: List<String>
-
   override suspend fun load(params: LoadParams<Int>): LoadResult<Int, UIPost> {
-    if (!::savedPosts.isInitialized) {
-      savedPosts = savedPostsRepository.savedPosts.first().map(SavedPost::shortId)
-    }
-    if (!::readPosts.isInitialized) {
-      readPosts = readPostsRepository.readPosts.first()
-    }
     val searchQuery = queryProvider()
     // If there is no query, we don't need to call the API at all.
     if (searchQuery.isEmpty()) {
@@ -66,15 +52,7 @@ constructor(
         val nextKey = if (result.value.isEmpty()) null else page + 1
         LoadResult.Page(
           itemsBefore = (page - 1) * PAGE_SIZE,
-          data =
-            result.value.map {
-              it
-                .toUIPost()
-                .copy(
-                  isSaved = savedPosts.contains(it.shortId),
-                  isRead = readPosts.contains(it.shortId),
-                )
-            },
+          data = result.value.map(LobstersPost::toUIPost),
           prevKey = if (page == STARTING_PAGE_INDEX) null else page - 1,
           nextKey = nextKey,
         )