+++ CHANGELOG.md
@@ -11,11 +11,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Opening posts you have previously seen will show the number of new comments since last visit
 
+### Fixed
+
+- Saving posts no longer triggers a page refresh that invalidates scroll position
+
 ### Changed
 
 - Change submitter text to 'authored' when applicable
 - Unread comments now have a brighter background rather than a text badge
-- Bottom navigation's visibility now interacts with the scroll behavior of the post lists
+- Bottom navigation bar has been redesigned
 
 ## [1.48.0] - 2024-06-05